pyproject.toml
[project]
name = "langchain-sidekick"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright",
    "langchain",
    "langchain-community",
    "langchain-experimental",
    "python-dotenv",
    "requests",
    "langgraph",
    "langchain-openai",
    "pydantic",
    "typing-extensions",
    "uuid",
    "gradio>=6.2.0",
    "beautifulsoup4",
    "wikipedia",
    "arxiv>=2.3.1",
    "yt-dlp>=2025.11.12",
    "lxml>=6.0.2",
    "langsmith>=0.4.45",
    "wolframalpha>=5.1.3",
    "wikibase-rest-api-client>=0.2.5",
    "mediawikiapi>=1.3",
    "langchain-tavily>=0.2.16",
    "ddgs>=9.10.0",
    "semanticscholar>=0.11.0",
    "youtube-search>=2.2.0",
    "google-search-results>=2.4.2",
    "pyowm>=3.5.0",
]
